Add multiple js bundles for different pages
