+++ package.json
@@ -12,6 +12,7 @@
     "build": "rollup -c"
   },
   "devDependencies": {
+    "globby": "^9.1.0",
     "rollup": "^1.4.1",
     "rollup-plugin-node-resolve": "^4.0.1"
   },
