+++ rollup.config.js
@@ -1,12 +1,17 @@
+// glob pattern input file matching from
+// https://github.com/rollup/rollup/issues/1435#issue-233487806
+
+import globby from 'globby';
+
 import nodeResolve from 'rollup-plugin-node-resolve';
 
-export default {
-    input: 'src/index.js',
+export default globby.sync('src/*.js').map(inputFile => ({
+    input: inputFile,
     output: {
-        file: 'dist/js/main.js',
+        file: inputFile.replace('src', 'dist/js'),
         format: 'iife',
     },
     plugins: [
         nodeResolve({ preferBuiltIns: true }),
     ],
-}
+}));
